+++ src/config.ts
@@ -2,4 +2,4 @@ const config = {
     basePath: 'https://kabbalahmedia.info',
     lang: 'en',
 };
-export default config
+export default config;
